refac: micro board files
