+++ lib/src/entities/micro_board/micro_board_handler.dart
@@ -1,0 +1,9 @@
+class MicroBoardHandler {
+  final String type;
+  final List<String> channels;
+
+  MicroBoardHandler({
+    required this.type,
+    required this.channels,
+  });
+}
