+++ lib/src/entities/micro_board/micro_board_route.dart
@@ -1,0 +1,9 @@
+class MicroBoardRoute {
+  final String route;
+  final String widget;
+
+  MicroBoardRoute({
+    required this.route,
+    required this.widget,
+  });
+}
