+++ lib/src/utils/constants/constants.dart
@@ -1,13 +1,15 @@
 class Constants {
   // Platform Channels
-  static String channelAppEvents = 'flutter/micro_app/app/events';
-  static String channelRouterLoggerEvents =
+  static const String channelAppEvents = 'flutter/micro_app/app/events';
+  static const String channelRouterLoggerEvents =
       'flutter/micro_app/router/logger_events';
-  static String channelRouterCommandEvents =
+  static const String channelRouterCommandEvents =
       'flutter/micro_app/router/command_events';
 
   // Methods
-  static String methodMicroAppEvent = 'app_event';
-  static String methodNavigationLog = 'navigation_log';
-  static String methodNavigationCommand = 'navigation_command';
+  static const String methodMicroAppEvent = 'app_event';
+  static const String methodNavigationLog = 'navigation_log';
+  static const String methodNavigationCommand = 'navigation_command';
+
+  static const String notTyped = 'Not Typed';
 }
